requirements.txt
azure_storage==0.37.0
deepgram_sdk==3.0.2
Flask==2.2.5
moviepy==1.0.3
nltk==3.8.1
pymongo==4.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.30.0
txtai==6.3.0
typing_extensions==4.10.0
